% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_api_key.R
\name{get_openai_api_key}
\alias{get_openai_api_key}
\alias{set_openai_api_key}
\title{Set and get Open AI API keys as ENV variable}
\usage{
get_openai_api_key()

set_openai_api_key(openai_api_key)
}
\arguments{
\item{openai_api_key}{Open AI API key/token to set as ENV}
}
\description{
Set and get Open AI API keys as ENV variable
}
