% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_chat.R
\name{openai_chat}
\alias{openai_chat}
\title{Open AI Chat API}
\usage{
openai_chat(
  prompt,
  openai_api_key = get_openai_api_key(),
  model = "gpt-3.5-turbo",
  chat_args = NULL
)
}
\arguments{
\item{openai_api_key}{An Open AI API key token.}

\item{model}{The model to use. See details below.}

\item{chat_args}{Further arguments to supply (excluding message and model).
See details below.}

\item{propt}{Character string to supply to Open AI chat.}
}
\description{
Open AI Chat API
}
\details{
Details on models can be found here:
\url{https://platform.openai.com/docs/models}

Open AIs chat documentation can be found here:
\url{https://platform.openai.com/docs/api-reference/chat}

Further arguments to the chat API can be found here:
\url{https://platform.openai.com/docs/api-reference/chat/create}
}
\examples{

\dontrun{
# Set the Open AI API token here:
set_open_ai_api_key("[YOUR KEY GOES HERE]")
# Test the API
openai_chat("Who is Gustav Vasa?")
# Test the API with specific arguments:
openai_chat("Who is Gustav Vasa?", chat_args = list(temperature = 0))
}
}
